chore: move main function to index file
